Add files via upload
This is the folder I used to run my node for question 2, it should be in the same folder as the rrbot_simulation_files folder. I'm also uploading an edited version of the rrbot.ros2_control.xacro file, as I needed to add the joint 3 information in order to get the information from every node from the /joint_states topic.